# v2/openapi/fb-unified-openapi.yaml
openapi: 3.0.0
servers:
  - description: Local server
    url: http://0.0.0.0:8000
info:
  title: Fireblocks Provider Connectivity API v2
  version: 0.5.0
  x-logo:
    url: doc-assets/fireblocks-logo-lockup-network-navy.png
    altText: Fireblocks logo
  contact:
    name: Fireblocks
    url: https://fireblocks.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  description: 
    $ref: README.md
...
paths:
...
  /capabilities/liquidity/quotes:
    get:
      operationId: getQuoteCapabilities
      summary: List possible asset conversions
      description: |
        Retrieves the list of supported asset conversion pairs that can be quoted. Shows which assets can be converted to other assets through the liquidity service.
      tags:
        - capabilities
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationStartingAfter'
        - $ref: '#/components/parameters/PaginationEndingBefore'
      responses:
        '200':
          description: List of possible asset conversions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteCapabilities'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
  /accounts/{accountId}/liquidity/quotes:
    post:
      operationId: createQuote
      summary: Quote creation request
      description: Either `fromAmount` or `toAmount` must be specified, while the unspecified amount will be populated by the server and returned in the response.
      tags:
        - liquidity
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      requestBody:
        description: Details of the quote request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequest'
      responses:
        '200':
          description: Quote details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
    get:
      operationId: getQuotes
      summary: Get list of quotes sorted by creation time
      description: |
        Retrieves a paginated list of all quotes for the specified account. Quotes are sorted by creation time and can be ordered ascending or descending.
      tags:
        - liquidity
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationStartingAfter'
        - $ref: '#/components/parameters/PaginationEndingBefore'
        - $ref: '#/components/parameters/ListOrderQueryParam'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          description: Quotes details.
          content:
            application/json:
              schema:
                type: object
                required:
                  - quotes
                properties:
                  quotes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Quote'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
  /accounts/{accountId}/liquidity/quotes/{id}:
    get:
      operationId: getQuoteDetails
      summary: Get quote details
      description: |
        Retrieves detailed information about a specific quote, including conversion rates, amounts, expiration time, and current status.
      tags:
        - liquidity
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/EntityIdPathParam'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          description: Quote details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
  /accounts/{accountId}/liquidity/quotes/{id}/execute:
    post:
      operationId: executeQuote
      summary: Execute quote
      description: |
        Executes a previously created quote, performing the actual asset conversion. The quote must be valid and not expired for execution to succeed.
      tags:
        - liquidity
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/EntityIdPathParam'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          description: Quote details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
...
  /accounts/{accountId}/capabilities/ramps:
    get:
      operationId: getRampMethods
      summary: Get list of supported ramp methods
      description: |
        Retrieves the list of supported on-ramp and off-ramp methods available for the specified account. Shows which payment methods and currencies are supported for fiat-to-crypto and crypto-to-fiat conversions.
      tags:
        - capabilities
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationStartingAfter'
        - $ref: '#/components/parameters/PaginationEndingBefore'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          $ref: '#/components/responses/RampMethodsResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
  /accounts/{accountId}/ramps:
    get:
      operationId: getRamps
      summary: Get list of ramps sorted by creation time
      description: |
        Retrieves a paginated list of all ramp transactions for the specified account. Ramps are sorted by creation time and include both on-ramp and off-ramp operations.
      tags:
        - ramps
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationStartingAfter'
        - $ref: '#/components/parameters/PaginationEndingBefore'
        - $ref: '#/components/parameters/ListOrderQueryParam'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          $ref: '#/components/responses/GetRampsResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
    post:
      operationId: createRamp
      summary: Initiates a new ramp
      description: Creates the payment instructions for the new ramp and return it in the response.
      tags:
        - ramps
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      requestBody:
        description: Ramp details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RampRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreateRampResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
  /accounts/{accountId}/ramps/{id}:
    get:
      operationId: getRampDetails
      summary: Get details of a specific ramp
      description: |
        Retrieves detailed information about a specific ramp transaction, including payment instructions, status, amounts, and processing details.
      tags:
        - ramps
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
        - $ref: '#/components/parameters/EntityIdPathParam'
      responses:
        '200':
          $ref: '#/components/responses/GetRampDetailsResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
components:
  parameters:
...
    X-FBAPI-KEY:
      name: X-FBAPI-KEY
      in: header
      required: true
      description: API authentication key.
      schema:
        type: string
    X-FBAPI-TIMESTAMP:
      name: X-FBAPI-TIMESTAMP
      in: header
      required: true
      description: Request timestamp in milliseconds since Unix epoch.
      schema:
        type: number
    X-FBAPI-NONCE:
      name: X-FBAPI-NONCE
      in: header
      required: true
      description: Unique identifier of the request.
      schema:
        type: string
    X-FBAPI-SIGNATURE:
      name: X-FBAPI-SIGNATURE
      in: header
      required: true
      description: |-
        Request signature using the chosen cryptographic algorithm. The signature is to be calculated on concatenation of the following request fields in the specified order:
        - `X-FBAPI-TIMESTAMP` - `X-FBAPI-NONCE` - HTTP request method in upper case - Endpoint path, including the query parameters - Request body
      schema:
        type: string
    X-FBAPI-INITIATED-BY:
      name: X-FBAPI-INITIATED-BY
      in: header
      required: false
      description: Conditional. This header is provided only when the corresponding capability requirement is enabled.
      schema:
        type: string
...
    EntityIdPathParam:
      name: id
      in: path
      required: true
      description: Entity unique identifier.
      schema:
        type: string
    SubAccountIdPathParam:
      name: accountId
      in: path
      required: true
      description: Sub-account identifier.
      schema:
        type: string
        pattern: ^[A-Za-z0-9_-]+$
    ListOrderQueryParam:
      name: order
      in: query
      description: Controls the order of the items in the returned list.
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc
    PaginationLimit:
      name: limit
      in: query
      description: Maximum number of returned items.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 10
    PaginationStartingAfter:
      name: startingAfter
      in: query
      description: Object ID. Instructs to return the items immediately following this object and not including it. Cannot be used together with `endingBefore`.
      schema:
        type: string
    PaginationEndingBefore:
      name: endingBefore
      in: query
      description: Object ID. Instructs to return the items immediately preceding this object and not including it. Cannot be used together with `startingAfter`.
      schema:
        type: string
...
  responses:
    BadRequestResponse:
      description: Request could not be processed due to a client error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestError'
    UnauthorizedResponse:
      description: Request is unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
    ErrorResponse:
      description: Failed to process request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GeneralError'
...
    RampMethodsResponse:
      description: List of ramp methods for account.
      content:
        application/json:
          schema:
            type: object
            required:
              - capabilities
            properties:
              capabilities:
                type: array
                items:
                  $ref: '#/components/schemas/RampMethod'
          examples:
            allUsdMethods:
              summary: USDs ramps
              description: |-
                This response indicates that on ramp from USD to USDC, off  ramp from USDC to USD and bridging between USDC and USDC on  different blockchains are supported. The USD asset can be  specified directly while tokens must be defined in  `GET /capabilities/assets` and referenced using their IDs.
                The asset IDs in this example are the same IDs used in the  examples of `GET /capabilities/assets`.
              value:
                capabilities:
                  - id: 16b8b2c3-bd61-4745-9c48-3d30c2bc6907
                    from:
                      asset:
                        nationalCurrencyCode: USD
                      transferMethod: Iban
                    to:
                      asset:
                        assetId: 360de0ad-9ba1-45d5-8074-22453f193d65
                      transferMethod: PublicBlockchain
                  - id: 69fa4f38-c1fd-4215-a310-82272225c4ef
                    from:
                      asset:
                        assetId: 360de0ad-9ba1-45d5-8074-22453f193d65
                      transferMethod: PublicBlockchain
                    to:
                      asset:
                        nationalCurrencyCode: USD
                      transferMethod: Iban
                  - id: afc7caa0-5b7e-4c98-8f2a-a037d0c6bbea
                    from:
                      asset:
                        assetId: 360de0ad-9ba1-45d5-8074-22453f193d65
                      transferMethod: PublicBlockchain
                    to:
                      asset:
                        assetId: 606bce6b-ff15-4704-9390-b9e32a6cfcff
                      transferMethod: PublicBlockchain
...
    GetRampDetailsResponse:
      description: Ramp details.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Ramp'
    GetRampsResponse:
      description: List of ramps for account.
      content:
        application/json:
          schema:
            type: object
            required:
              - ramps
            properties:
              ramps:
                type: array
                items:
                  $ref: '#/components/schemas/Ramp'
          examples:
            allRamps:
              summary: All ramps
              description: This response indicates that there are 3 ramps in the account.
              value:
                ramps:
                  - id: 16b8b2c3-bd61-4745-9c48-3d30c2bc6907
                    createdAt: '2021-09-01T12:00:00Z'
                    updatedAt: '2021-09-01T12:00:00Z'
                    status: Pending
                    amount: '1231.45'
                    type: OnRamp
                    from:
                      asset:
                        nationalCurrencyCode: USD
                      transferMethod: Iban
                    to:
                      asset:
                        assetId: 360de0ad-9ba1-45d5-8074-22453f193d65
                      transferMethod: PublicBlockchain
                      address: '0x1234567890abcdef1234567890abcdef12345678'
                      tag: 1234567890abcdef
                    paymentInstructions:
                      asset:
                        nationalCurrencyCode: USD
                      transferMethod: Iban
                      accountHolder:
                        name: John Doe
                        address: 123 Main St, New York, NY 10001
                      iban: US1234567890
                    estimatedFees:
                      - feeType: ORDER
                        feeAsset:
                          nationalCurrencyCode: USD
                        fee:
                          amountType: FIXED
                          amount: '100.00'
                    expiresAt: '2021-09-01T12:00:00Z'
                  - id: 69fa4f38-c1fd-4215-a310-82272225c4ef
                    createdAt: '2021-09-01T12:00:00Z'
                    updatedAt: '2021-09-01T12:00:00Z'
                    status: Completed
                    amount: '1231.45'
                    type: OffRamp
                    from:
                      asset:
                        assetId: 360de0ad-9ba1-45d5-8074-22453f193d65
                      transferMethod: PublicBlockchain
                    to:
                      asset:
                        nationalCurrencyCode: USD
                      transferMethod: Iban
                      accountHolder:
                        name: John Doe
                        address: 123 Main St, New York, NY 10001
                      iban: US1234567890
                    paymentInstructions:
                      asset:
                        assetId: 360de0ad-9ba1-45d5-8074-22453f193d65
                      transferMethod: PublicBlockchain
                      address: '0x1234567890abcdef1234567890abcdef12345678'
                    payment:
                      blockchainTxId: '0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef'
                      amount: '1231.45'
                      address: '0x1234567890abcdef1234567890abcdef12345678'
                      asset:
                        assetId: 360de0ad-9ba1-45d5-8074-22453f193d65
                      transferMethod: PublicBlockchain
                    receipt:
                      asset:
                        nationalCurrencyCode: USD
                      transferMethod: Iban
                      accountHolder:
                        name: John Doe
                        address: 123 Main St, New York, NY 10001
                      iban: US1234567890
                      amount: '1231.45'
                      actualFees:
                        - feeType: ORDER
                          feeAsset:
                            nationalCurrencyCode: USD
                          fee:
                            amountType: FIXED
                            amount: '100.00'
                    expiresAt: '2021-09-01T12:00:00Z'
                  - id: afc7caa0-5b7e-4c98-8f2a-a037d0c6bbea
                    createdAt: '2021-09-01T12:00:00Z'
                    updatedAt: '2021-09-01T12:00:00Z'
                    status: Processing
                    amount: '1231.45'
                    type: Bridge
                    from:
                      asset:
                        assetId: 360de0ad-9ba1-45d5-8074-22453f193d65
                      transferMethod: PublicBlockchain
                    to:
                      asset:
                        assetId: 606bce6b-ff15-4704-9390-b9e32a6cfcff
                      transferMethod: PublicBlockchain
                      address: '0x1234567890abcdef1234567890abcdef12345678'
                      tag: 1234567890abcdef
                    paymentInstructions:
                      asset:
                        assetId: 360de0ad-9ba1-45d5-8074-22453f193d65
                      transferMethod: PublicBlockchain
                      address: '0x1234567890abcdef1234567890abcdef12355678'
                    payment:
                      blockchainTxId: '0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef'
                      amount: '1231.45'
                      address: '0x1234567890abcdef1234567890abcdef12355678'
                      asset:
                        assetId: 360de0ad-9ba1-45d5-8074-22453f193d65
                      transferMethod: PublicBlockchain
                    receipt:
                      actualFees:
                        - feeType: ORDER
                          feeAsset:
                            nationalCurrencyCode: USD
                          fee:
                            amountType: FIXED
                            amount: '100.00'
                      blockchainTxId: '0x1234567890abcdef1234567899abcdef1234567890abcdef1234567890abcdef'
                      amount: '1231.45'
                      address: '0x1234567890abcdef1234567890abcdef12345678'
                      asset:
                        assetId: 606bce6b-ff15-4704-9390-b9e32a6cfcff
                      transferMethod: PublicBlockchain
                    estimatedFees:
                      - feeType: ORDER
                        feeAsset:
                          nationalCurrencyCode: USD
                        fee:
                          amountType: FIXED
                          amount: '100.00'
                    expiresAt: '2021-09-01T12:00:00Z'
  schemas:
...
    PositiveAmount:
      type: string
      pattern: ^\d+(\.\d+)?
...
    QuoteCapabilities:
      type: object
      required:
        - capabilities
      properties:
        capabilities:
          type: array
          items:
            $ref: '#/components/schemas/QuoteCapability'
...
    QuoteRequest:
      type: object
      allOf:
        - type: object
          required:
            - fromAsset
            - toAsset
          properties:
            fromAsset:
              $ref: '#/components/schemas/AssetReference'
            toAsset:
              $ref: '#/components/schemas/AssetReference'
        - oneOf:
            - type: object
              required:
                - fromAmount
              properties:
                fromAmount:
                  $ref: '#/components/schemas/PositiveAmount'
            - type: object
              required:
                - toAmount
              properties:
                toAmount:
                  $ref: '#/components/schemas/PositiveAmount'
    Quote:
      type: object
      required:
        - id
        - fromAsset
        - fromAmount
        - toAsset
        - toAmount
        - conversionFeeBps
        - status
        - createdAt
        - expiresAt
      properties:
        id:
          type: string
        fromAsset:
          $ref: '#/components/schemas/AssetReference'
        fromAmount:
          $ref: '#/components/schemas/PositiveAmount'
        toAsset:
          $ref: '#/components/schemas/AssetReference'
        toAmount:
          $ref: '#/components/schemas/PositiveAmount'
        conversionFeeBps:
          type: integer
          description: Conversion fee in basis points. The amounts are expected to include the fee.
        conversionFeeAsset:
          allOf:
            - $ref: '#/components/schemas/AssetReference'
            - description: |
                This is the asset that will be used to pay the conversion fee. If not returned, the asset will be the same as the fromAsset.
        status:
          $ref: '#/components/schemas/QuoteStatus'
        createdAt:
          description: Quote creation time.
          type: string
          format: date-time
        expiresAt:
          description: Quote expiration time.
          type: string
          format: date-time
...
  securitySchemes: {}